add condition on admincontroller and edit readme.md

+++ app/Http/Controllers/AdminController.php
@@ -16,37 +16,55 @@ class AdminController extends Controller
      */
     public function showAllUsers(Request $request)
     {
-        $user = User::all();
-        if ($user) {
-            return response()->json([
-                'status' => true,
-                'message' => 'get all users successfully',
-                'data' => $user
-            ], 200);
+        $user_id = User::where('api_token', $request->api_token)->first();
+        if ($user_id->role == 'admin') {
+            $user = User::all();
+            if ($user) {
+                return response()->json([
+                    'status' => true,
+                    'message' => 'get all users successfully',
+                    'data' => $user
+                ], 200);
+            } else {
+                return response()->json([
+                    'status' => false,
+                    'message' => 'failed. data not found',
+                    'data' => null
+                ], 404);
+            }
         } else {
             return response()->json([
                 'status' => false,
-                'message' => 'failed. data not found',
+                'message' => 'failed. forbidden to access',
                 'data' => null
-            ], 404);
+            ], 403);
         }
     }
 
     public function showDetailUser(Request $request, $id)
     {
-        $user = User::find($id);
-        if ($user) {
-            return response()->json([
-                'status' => true,
-                'message' => 'get detail user successfully',
-                'data' => $user
-            ], 200);
+        $user_id = User::where('api_token', $request->api_token)->first();
+        if ($user_id->role == 'admin') {
+            $user = User::find($id);
+            if ($user) {
+                return response()->json([
+                    'status' => true,
+                    'message' => 'get detail user successfully',
+                    'data' => $user
+                ], 200);
+            } else {
+                return response()->json([
+                    'status' => false,
+                    'message' => 'failed. data not found',
+                    'data' => null
+                ], 404);
+            }
         } else {
             return response()->json([
                 'status' => false,
-                'message' => 'failed. data not found',
+                'message' => 'failed. forbidden to access',
                 'data' => null
-            ], 404);
+            ], 403);
         }
     }
 
@@ -62,22 +80,86 @@ class AdminController extends Controller
                 'message' => $validator->errors()
             ], 400);
         } else {
-            $user = User::find($id);
-            if ($user) {
-                $data = $user->update([
-                    'role' => $request->role
-                ]);
-                if ($data) {
-                    return response()->json([
-                        'status' => true,
-                        'message' => 'update successfully',
-                    ], 201);
+            $user_id = User::where('api_token', $request->api_token)->first();
+            if ($user_id->role == 'admin') {
+                $user = User::find($id);
+                if ($user) {
+                    $data = $user->update([
+                        'role' => $request->role
+                    ]);
+                    if ($data) {
+                        return response()->json([
+                            'status' => true,
+                            'message' => 'update successfully',
+                        ], 201);
+                    } else {
+                        return response()->json([
+                            'status' => false,
+                            'message' => 'update failed!',
+                        ], 400);
+                    }
                 } else {
                     return response()->json([
                         'status' => false,
-                        'message' => 'update failed!',
-                    ], 400);
+                        'message' => 'failed. data not found',
+                        'data' => null
+                    ], 404);
                 }
+            } else {
+                return response()->json([
+                    'status' => false,
+                    'message' => 'failed. forbidden to access',
+                    'data' => null
+                ], 403);
+            }
+        }
+    }
+
+    public function deleteUser(Request $request, $id)
+    {
+        $user_id = User::where('api_token', $request->api_token)->first();
+        if ($user_id->role == 'admin') {
+            $user = User::find($id);
+
+            if ($user) {
+                if ($user->picture) {
+                    $path_name = 'users_profile/' . $user->picture;
+                    if (file_exists($path_name)) {
+                        unlink($path_name);
+                    }
+                }
+                $user->delete();
+
+                return response()->json([
+                    'status' => true,
+                    'message' => 'delete successfully',
+                ], 200);
+            } else {
+                return response()->json([
+                    'status' => false,
+                    'message' => 'delete failed. Data not found',
+                ], 404);
+            }
+        } else {
+            return response()->json([
+                'status' => false,
+                'message' => 'failed. forbidden to access',
+                'data' => null
+            ], 403);
+        }
+    }
+
+    public function showAllEvents(Request $request)
+    {
+        $user_id = User::where('api_token', $request->api_token)->first();
+        if ($user_id->role == 'admin') {
+            $event = Event::with(['user', 'category'])->get();
+            if ($event) {
+                return response()->json([
+                    'status' => true,
+                    'message' => 'get all events successfully',
+                    'data' => $event
+                ], 200);
             } else {
                 return response()->json([
                     'status' => false,
@@ -85,49 +167,12 @@ class AdminController extends Controller
                     'data' => null
                 ], 404);
             }
-        }
-    }
-
-    public function deleteUser($id)
-    {
-        $user = User::find($id);
-
-        if ($user) {
-            if ($user->picture) {
-                $path_name = 'users_profile/' . $user->picture;
-                if (file_exists($path_name)) {
-                    unlink($path_name);
-                }
-            }
-            $user->delete();
-
-            return response()->json([
-                'status' => true,
-                'message' => 'delete successfully',
-            ], 200);
         } else {
             return response()->json([
                 'status' => false,
-                'message' => 'delete failed. Data not found',
-            ], 404);
-        }
-    }
-
-    public function showAllEvents()
-    {
-        $event = Event::with(['user', 'category'])->get();
-        if ($event) {
-            return response()->json([
-                'status' => true,
-                'message' => 'get all events successfully',
-                'data' => $event
-            ], 200);
-        } else {
-            return response()->json([
-                'status' => false,
-                'message' => 'failed. data not found',
+                'message' => 'failed. forbidden to access',
                 'data' => null
-            ], 404);
+            ], 403);
         }
     }
 
@@ -143,54 +188,72 @@ class AdminController extends Controller
                 'message' => $validator->errors()
             ], 400);
         } else {
-            $event = Event::find($id);
-            if ($event) {
-                $data = $event->update([
-                    'status' => $request->status
-                ]);
-                if ($data) {
-                    return response()->json([
-                        'status' => true,
-                        'message' => 'update successfully',
-                    ], 200);
+            $user_id = User::where('api_token', $request->api_token)->first();
+            if ($user_id->role == 'admin') {
+                $event = Event::find($id);
+                if ($event) {
+                    $data = $event->update([
+                        'status' => $request->status
+                    ]);
+                    if ($data) {
+                        return response()->json([
+                            'status' => true,
+                            'message' => 'update successfully',
+                        ], 200);
+                    } else {
+                        return response()->json([
+                            'status' => false,
+                            'message' => 'update failed!',
+                        ], 400);
+                    }
                 } else {
                     return response()->json([
                         'status' => false,
-                        'message' => 'update failed!',
-                    ], 400);
+                        'message' => 'failed. data not found',
+                        'data' => null
+                    ], 404);
                 }
             } else {
                 return response()->json([
                     'status' => false,
-                    'message' => 'failed. data not found',
+                    'message' => 'failed. forbidden to access',
                     'data' => null
-                ], 404);
+                ], 403);
             }
         }
     }
 
-    public function deleteEvent($id)
+    public function deleteEvent(Request $request, $id)
     {
-        $event = Event::find($id);
+        $user_id = User::where('api_token', $request->api_token)->first();
+        if ($user_id->role == 'admin') {
+            $event = Event::find($id);
 
-        if ($event) {
-            if ($event->poster) {
-                $path_name = 'poster_uploads/' . $event->poster;
-                if (file_exists($path_name)) {
-                    unlink($path_name);
+            if ($event) {
+                if ($event->poster) {
+                    $path_name = 'poster_uploads/' . $event->poster;
+                    if (file_exists($path_name)) {
+                        unlink($path_name);
+                    }
                 }
-            }
-            $event->delete();
+                $event->delete();
 
-            return response()->json([
-                'status' => true,
-                'message' => 'delete successfully',
-            ], 200);
+                return response()->json([
+                    'status' => true,
+                    'message' => 'delete successfully',
+                ], 200);
+            } else {
+                return response()->json([
+                    'status' => false,
+                    'message' => 'delete failed. Data not found',
+                ], 404);
+            }
         } else {
             return response()->json([
                 'status' => false,
-                'message' => 'delete failed. Data not found',
-            ], 404);
+                'message' => 'failed. forbidden to access',
+                'data' => null
+            ], 403);
         }
     }
